requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
anyio==3.6.2
    # via jupyter-server
argon2-cffi==21.3.0
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==2.2.1
    # via stack-data
attrs==23.1.0
    # via jsonschema
babel==2.12.1
    # via jupyterlab-server
backcall==0.2.0
    # via ipython
beautifulsoup4==4.12.2
    # via nbconvert
bleach==6.0.0
    # via nbconvert
build==0.10.0
    # via pip-tools
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via argon2-cffi-bindings
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via pip-tools
comm==0.1.3
    # via ipykernel
debugpy==1.6.7
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
executing==1.2.0
    # via stack-data
fastjsonschema==2.16.3
    # via nbformat
idna==3.4
    # via
    #   anyio
    #   requests
importlib-metadata==6.6.0
    # via
    #   jupyter-client
    #   jupyterlab-server
    #   nbconvert
intelhex==2.3.0
    # via sensirion-shdlc-driver
ipykernel==6.22.0
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   nbclassic
    #   notebook
    #   qtconsole
ipython==8.13.2
    # via
    #   -r requirements.in
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
    #   jupyterlab
ipython-genutils==0.2.0
    # via
    #   nbclassic
    #   notebook
    #   qtconsole
ipywidgets==8.0.6
    # via jupyter
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbclassic
    #   nbconvert
    #   notebook
json5==0.9.11
    # via jupyterlab-server
jsonschema==4.17.3
    # via
    #   jupyterlab-server
    #   nbformat
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==8.2.0
    # via
    #   ipykernel
    #   jupyter-console
    #   jupyter-server
    #   nbclassic
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.6.3
    # via jupyter
jupyter-core==5.3.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   jupyterlab
    #   nbclassic
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyter-server==1.24.0
    # via
    #   jupyterlab
    #   jupyterlab-server
    #   nbclassic
    #   notebook-shim
jupyterlab==3.5.3
    # via -r requirements.in
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-server==2.15.2
    # via
    #   -r requirements.in
    #   jupyterlab
jupyterlab-widgets==3.0.7
    # via ipywidgets
markupsafe==2.1.2
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mistune==2.0.5
    # via nbconvert
mysql-connector-python==8.0.33
    # via -r requirements.in
nbclassic==1.0.0
    # via
    #   jupyterlab
    #   notebook
nbclient==0.7.4
    # via nbconvert
nbconvert==7.3.1
    # via
    #   jupyter
    #   jupyter-server
    #   nbclassic
    #   notebook
nbformat==5.8.0
    # via
    #   jupyter-server
    #   nbclassic
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.6
    # via
    #   ipykernel
    #   nbclassic
    #   notebook
notebook==6.5.4
    # via
    #   jupyter
    #   jupyterlab
notebook-shim==0.2.3
    # via nbclassic
packaging==23.1
    # via
    #   build
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   qtconsole
    #   qtpy
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.13.0
    # via -r requirements.in
platformdirs==3.5.0
    # via jupyter-core
prometheus-client==0.16.0
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
prompt-toolkit==3.0.38
    # via
    #   ipython
    #   jupyter-console
protobuf==3.20.3
    # via mysql-connector-python
psutil==5.9.5
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.15.1
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   qtconsole
pyproject-hooks==1.0.0
    # via build
pyrsistent==0.19.3
    # via jsonschema
pyserial==3.5
    # via sensirion-shdlc-driver
python-dateutil==2.8.2
    # via jupyter-client
pyzmq==25.0.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   nbclassic
    #   notebook
    #   qtconsole
qtconsole==5.4.3
    # via jupyter
qtpy==2.3.1
    # via qtconsole
requests==2.30.0
    # via jupyterlab-server
send2trash==1.8.2
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
sensirion-i2c-driver==1.0.0
    # via sensirion-i2c-scd
sensirion-i2c-scd==0.1.2
    # via -r requirements.in
sensirion-shdlc-driver[fwupdate]==0.1.5
    # via sensirion-shdlc-sensorbridge
sensirion-shdlc-sensorbridge==0.1.1
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
sniffio==1.3.0
    # via anyio
soupsieve==2.4.1
    # via beautifulsoup4
stack-data==0.6.2
    # via ipython
terminado==0.17.1
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
tinycss2==1.2.1
    # via nbconvert
tomli==2.0.1
    # via
    #   build
    #   jupyterlab
    #   pyproject-hooks
tornado==6.3.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbclassic
    #   notebook
    #   terminado
traitlets==5.9.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-console
    #   jupyter-core
    #   jupyter-server
    #   matplotlib-inline
    #   nbclassic
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
typing-extensions==4.5.0
    # via ipython
urllib3==2.0.2
    # via requests
wcwidth==0.2.6
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.5.1
    # via jupyter-server
wheel==0.40.0
    # via pip-tools
widgetsnbextension==4.0.7
    # via ipywidgets
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
